separate search into separate component to reduce re-renders
